Setup radians/degrees key
This keys changes the unit of measurement for angles to either radians or degrees.  The display area has an indicator that shows the current unit. The key's label also changes based on the current setting, ie when unit is set to degrees, the key is labeled "Rad" meaning change unit to radians.